+++ lib/fulfil_api/configuration.rb
@@ -7,7 +7,7 @@ module FulfilApi
   #   to these settings.
   class Configuration
     attr_accessor :access_token, :api_version, :merchant_id, :request_options, :tpl
-    attr_reader :connection_options
+    attr_reader :connection_options, :oauth
 
     DEFAULT_API_VERSION = "v2"
     DEFAULT_REQUEST_OPTIONS = { open_timeout: 1, read_timeout: 5, write_timeout: 5, timeout: 5 }.freeze
@@ -48,7 +48,33 @@ module FulfilApi
       @connection_options = DEFAULT_CONNECTION_OPTIONS.merge(options || {})
     end
 
+    # Assigns the credentials of the OAuth app that was registered in Fulfil's
+    #   authentication dashboard. Assigning `nil` resets to the defaults.
+    #
+    # @param options [Hash, FulfilApi::OAuth::Configuration, nil] The OAuth options to apply.
+    # @return [void]
+    def oauth=(options)
+      @oauth =
+        case options
+        when OAuth::Configuration then options
+        when Hash, NilClass then OAuth::Configuration.new(options)
+        else raise ArgumentError, "Expected Hash or FulfilApi::OAuth::Configuration, got #{options.class} instead"
+        end
+    end
+
     private
+
+    # {FulfilApi.with_config} duplicates the active configuration, which would
+    #   otherwise share the very same OAuth configuration object with the copy.
+    #   A block that changes an OAuth option would then leak that change into
+    #   the configuration it was supposed to be reverted to.
+    #
+    # @param source [FulfilApi::Configuration] The configuration being copied.
+    # @return [void]
+    def initialize_copy(source)
+      super
+      @oauth = source.oauth.dup
+    end
 
     # Sets the default options for the gem configuration.
     #
@@ -60,6 +86,7 @@ module FulfilApi
       self.api_version = DEFAULT_API_VERSION if api_version.nil?
       self.request_options = DEFAULT_REQUEST_OPTIONS if request_options.nil?
       self.connection_options = nil if connection_options.nil?
+      self.oauth = nil if oauth.nil?
     end
   end